requirements.txt
uvicorn[standard] == 0.20.0
fastapi == 0.111.0
loguru
PyJWT
bcrypt
blinker
ujson
pydantic == 2.7.4
pydantic_settings ==2.5.2
pandas
python-multipart
sqlalchemy[asyncio] == 2.0.36
asyncpg
python-jose[cryptography]
# azure-storage-blob >= 12.20.0
# azure-identity >= 1.17.1
# langfuse
# chromadb
# openai
dateparser >= 1.2.0
# pdfminer-six == 20231228
sqlalchemy-filterset >= 2.0.0
# apscheduler >= 3.10.4
regex >= 2024.5.15
# fastapi_mail
pylint
# langchain
# langchain-community
# langchain-core
# langchain-openai
# langchain-chroma
# phonenumbers
# pymupdf4llm
# markdown
# weasyprint
# fastapi_mail
